+++ packages/create-onchain/templates/minikit-snake/app/components/snake.tsx
@@ -46,7 +46,6 @@ import {
 } from "@coinbase/onchainkit/transaction";
 import {
   ConnectWallet,
-  ConnectWalletText,
   Wallet,
   WalletDropdown,
   WalletDropdownDisconnect,
@@ -344,8 +343,10 @@ function ControlButton({ children, onClick, className }: ControlButtonProps) {
 function WalletControl() {
   return (
     <Wallet className="[&>div:nth-child(2)]:!opacity-20 md:[&>div:nth-child(2)]:!opacity-100">
-      <ConnectWallet className="w-12 h-12 bg-[#0052FF] rounded-full hover:bg-[#0052FF] focus:bg-[#0052FF] cursor-pointer select-none transition-all duration-150 border-[1px] border-[#0052FF] min-w-12 [box-shadow:0_5px_0_0_#002299,0_8px_0_0_#0033cc33]">
-        <ConnectWalletText>{""}</ConnectWalletText>
+      <ConnectWallet
+        className="w-12 h-12 bg-[#0052FF] rounded-full hover:bg-[#0052FF] focus:bg-[#0052FF] cursor-pointer select-none transition-all duration-150 border-[1px] border-[#0052FF] min-w-12 [box-shadow:0_5px_0_0_#002299,0_8px_0_0_#0033cc33]"
+        disconnectedLabel={' '}
+      >{' '}
       </ConnectWallet>
       <WalletDropdown>
         <Identity className="px-4 pt-3 pb-2" hasCopyAddressOnClick>
@@ -531,11 +532,7 @@ export function Dead({ score, level, onGoToIntro, isWin }: DeadProps) {
   const transactionButton = useMemo(() => {
     if (!address) {
       return (
-        <Wallet>
-          <ConnectWallet>
-            <ConnectWalletText>Login to save your high score</ConnectWalletText>
-          </ConnectWallet>
-        </Wallet>
+        <ConnectWallet disconnectedLabel="Login to save your high score"/>
       );
     }